pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "emailbison"
version = "0.1.0"
description = "CLI tool to interact with the EmailBison API"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Brandon Pettee" }]
dependencies = [
  "typer>=0.12.3",
  "httpx>=0.27.0",
  "pydantic>=2.6",
  "platformdirs>=4.2",
  "python-dateutil>=2.9",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "respx>=0.21.1",
  "ruff>=0.4.0",
]

[project.scripts]
emailbison = "emailbison.cli:app"
eb = "emailbison.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["emailbison"]
